requirements.txt
pix2tex>=0.1.4
psycopg2-binary>=2.9.9
Pillow>=10.0.0
matplotlib>=3.8.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
